pyproject.toml
[project]
name = "algorithmics-traiing-strategies"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
authors = [
    { name = "Dr. Alfred Ang", email = "angch@tertiaryinfotech.com" }
]
requires-python = ">=3.13"
dependencies = [
    "futu-api>=9.6.5608",
    "google-generativeai>=0.8.6",
    "matplotlib>=3.10.8",
    "numpy>=2.4.2",
    "pandas>=2.3.3",
    "plotly>=6.5.2",
    "python-dotenv>=1.2.1",
    "pyyaml>=6.0.3",
    "setuptools>=82.0.0",
    "streamlit>=1.54.0",
    "tqdm>=4.67.3",
    "yfinance>=1.1.0",
]

[tool.hatch.build.targets.wheel]
packages = ["src"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
